requirements.txt
discord.py==2.4.0
rich==13.7.1
